gmx rotmat support and direct transform support
Included an option to calculate the global rotational diffusion from the rotational matrices given by gmx rotmat. However, this is not quite the same as the PLUMED results probably due to the precision at which the quaternions and the rotational matrices are computed and stored.
Separately, added support to directly compute relaxation from fitted-C(t) values. This is now the default behaviour when no arguments w.r.t. global rotational diffusion is given to calculate-relaxations-from-Ct.py.