Bugfixes and optimisations part 3 - histograms and name changes
Given that this workflow is meant to work with large trajectories, made histograms the default action to store the X-H vector distributions converted from the trajectory.
This induces a small error but at a vastly improved performance.
Name changes to how the files are stored are done to help specify the various options.
Also started an class rewrite of relaxModel -> relaxMolecule for future support of multiple diffusions tensors in the same molecule.