Code Rewrite for C(t) modules - fitting and plotting
Introduced the fitParams class for properly handling the internal
autocorrelations of spin relaxation. This will be renamed again in the
next edit.

The curve-fitting calculations and plotting of C(t) functions have been
adapted to use this Class. Compute-relaxations will be converted
next to simplify the code. This wil hopefully open the way for
multi-experiment handling.