Second Stage prep for fitting CSA variations and multifield experiments
Contains several bug-fixes and splitting up the rho computation.

A new class and subclasses have beenj created to handle all components
associated with global rotational diffusion. This is a new rotdifModel
that needs to both handle tensor D as well as bond vector definitions,
and is self-sufficient to generate J(omega,autoCorr)